feat: added Latex translations via Katex! 😎
The latex conversion process detects $...$ and $$...$$ patterns appropriately,
and ignores them when they are between code blocks.

This isn't 100% clean yet, as the katex css file is taken from the web,
instead of compiled locally. This will change in the future, with rollup config file
written appropriately.

It should be possible to pass some custom macros to Katex when a "preferences.json" file
will be available.