feat: added Github Markdown stylesheet + highlight.js syntax highlighting
This is a needed feature, but I'm afraid it slows down the app reactivity.

Having now markdown-it + katex + highlight.js rendering in real time is somehow a lot.

In addition, there is more to come: CodeJar or similar to render the leftpane.

I might find something to render less often, instead of rendering every time
a single character changed from the source code.