feat: dynamically compute optimal interval between renders
This computes a more accurate number than a hardcoded constant.
When having small documents, re-rendering every 5ms is largely sufficient (and user will have a real-time feeling).
Doing the same for large documents become slower and slower, because the render time becomes significant.

I opted for a dynamical way of computing the optimal interval between two render passes, according to the last 10 render times.

The interval time is computed as follow: 'mean + 110% of std_deviation'
This gives a decent value most of the time, going more and more optimal as new renders are done (very rapidly, 10 renders are reached very fast).

I first tried with 50 but it seems to be unecessary, this may change in the future. Let's Keep It Simple for now.